requirements.txt
# Core dependencies
firecrawl-py>=1.0.0
resend>=2.0.0
pymongo>=4.6.0

# Type checking & validation
pydantic>=2.0.0

# Environment & config
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# Dev dependencies (optional)
pytest>=8.0.0
